Refactor YAxis and ZAxis configurations in safety
